wip split website quote
