+++ addons/sale/__manifest__.py
@@ -27,6 +27,7 @@ This module contains all the common features of Sales Management and eCommerce.
         'views/mail_activity_views.xml',
         'views/assets.xml',
         'views/sale_portal_templates.xml',
+        'views/sale_quote_views.xml',
         'views/sale_onboarding_views.xml',
         'views/res_config_settings_views.xml',
         'views/res_users_views.xml',
