+++ addons/sale/models/sale_quote_line.py
@@ -1,0 +1,74 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from odoo import api, fields, models
+from odoo.addons import decimal_precision as dp
+from odoo.exceptions import UserError
+
+
+class SaleQuoteLine(models.Model):
+    _name = "sale.quote.line"
+    _description = "Quotation Template Lines"
+    _order = 'quote_id, sequence, id'
+
+    sequence = fields.Integer('Sequence', help="Gives the sequence order when displaying a list of sale quote lines.",
+        default=10)
+    quote_id = fields.Many2one('sale.quote.template', 'Quotation Template Reference', required=True,
+        ondelete='cascade', index=True)
+    name = fields.Text('Description', required=True, translate=True)
+    product_id = fields.Many2one('product.product', 'Product', domain=[('sale_ok', '=', True)])
+    price_unit = fields.Float('Unit Price', required=True, digits=dp.get_precision('Product Price'))
+    discount = fields.Float('Discount (%)', digits=dp.get_precision('Discount'), default=0.0)
+    product_uom_qty = fields.Float('Quantity', required=True, digits=dp.get_precision('Product UoS'), default=1)
+    product_uom_id = fields.Many2one('uom.uom', 'Unit of Measure')
+
+    display_type = fields.Selection([
+        ('line_section', "Section"),
+        ('line_note', "Note")], default=False, help="Technical field for UX purpose.")
+
+    @api.onchange('product_id')
+    def _onchange_product_id(self):
+        self.ensure_one()
+        if self.product_id:
+            name = self.product_id.name_get()[0][1]
+            if self.product_id.description_sale:
+                name += '\n' + self.product_id.description_sale
+            self.name = name
+            self.price_unit = self.product_id.lst_price
+            self.product_uom_id = self.product_id.uom_id.id
+            domain = {'product_uom_id': [('category_id', '=', self.product_id.uom_id.category_id.id)]}
+            return {'domain': domain}
+
+    @api.onchange('product_uom_id')
+    def _onchange_product_uom(self):
+        if self.product_id and self.product_uom_id:
+            self.price_unit = self.product_id.uom_id._compute_price(self.product_id.lst_price, self.product_uom_id)
+
+    @api.model
+    def create(self, values):
+        if values.get('display_type', self.default_get(['display_type'])['display_type']):
+            values.update(product_id=False, price_unit=0, product_uom_qty=0, product_uom_id=False)
+        return super(SaleQuoteLine, self).create(values)
+
+    @api.multi
+    def write(self, values):
+        if 'display_type' in values and self.filtered(lambda line: line.display_type != values.get('display_type')):
+            raise UserError("You cannot change the type of a sale quote line. Instead you should delete the current line and create a new line of the proper type.")
+        return super(SaleQuoteLine, self).write(values)
+
+    _sql_constraints = [
+        ('accountable_product_id_required',
+            "CHECK(display_type IS NOT NULL OR (product_id IS NOT NULL AND product_uom_id IS NOT NULL))",
+            "Missing required product and UoM on accountable sale quote line."),
+
+        ('non_accountable_fields_null',
+            "CHECK(display_type IS NULL OR (product_id IS NULL AND price_unit = 0 AND product_uom_qty = 0 AND product_uom_id IS NULL))",
+            "Forbidden product, unit price, quantity, and UoM on non-accountable sale quote line"),
+    ]
+
+    def _get_vals_to_apply_template(self):
+        return {
+            'display_type': self.display_type,
+            'name': self.name,
+            'state': 'draft',
+        }
