+++ addons/sale/models/sale_quote_option.py
@@ -1,0 +1,49 @@
+# -*- coding: utf-8 -*-
+# Part of Odoo. See LICENSE file for full copyright and licensing details.
+
+from odoo import api, fields, models
+from odoo.addons import decimal_precision as dp
+
+
+class SaleQuoteOption(models.Model):
+    _name = "sale.quote.option"
+    _description = "Quotation Option"
+
+    template_id = fields.Many2one('sale.quote.template', 'Quotation Template Reference', ondelete='cascade',
+        index=True, required=True)
+    name = fields.Text('Description', required=True, translate=True)
+    product_id = fields.Many2one('product.product', 'Product', domain=[('sale_ok', '=', True)], required=True)
+    price_unit = fields.Float('Unit Price', required=True, digits=dp.get_precision('Product Price'))
+    discount = fields.Float('Discount (%)', digits=dp.get_precision('Discount'))
+    uom_id = fields.Many2one('uom.uom', 'Unit of Measure ', required=True)
+    quantity = fields.Float('Quantity', required=True, digits=dp.get_precision('Product UoS'), default=1)
+
+    @api.onchange('product_id')
+    def _onchange_product_id(self):
+        if not self.product_id:
+            return
+        self.price_unit = self.product_id.list_price
+        self.name = self.product_id.name
+        self.uom_id = self.product_id.uom_id
+        domain = {'uom_id': [('category_id', '=', self.product_id.uom_id.category_id.id)]}
+        return {'domain': domain}
+
+    @api.onchange('uom_id')
+    def _onchange_product_uom(self):
+        if not self.product_id:
+            return
+        if not self.uom_id:
+            self.price_unit = 0.0
+            return
+        if self.uom_id.id != self.product_id.uom_id.id:
+            self.price_unit = self.product_id.uom_id._compute_price(self.price_unit, self.uom_id)
+
+    def _get_vals_to_apply_template(self):
+        return {
+            'product_id': self.product_id.id,
+            'name': self.name,
+            'quantity': self.quantity,
+            'uom_id': self.uom_id.id,
+            'price_unit': self.price_unit,
+            'discount': self.discount,
+        }
