+++ addons/sale/views/sale_views.xml
@@ -215,6 +215,9 @@
                     <button name="action_confirm"
                         string="Confirm Sale" type="object"
                         attrs="{'invisible': [('state', 'not in', ['draft'])]}"/>
+                    <button name="open_quotation" type="object"
+                        class="o_open_quotation" string="Preview"
+                        attrs="{'invisible': [('template_id', '=', False)]}"/>
                     <button name="action_quotation_send" type="object" string="Send PRO-FORMA Invoice" groups="sale.group_proforma_sales" attrs="{'invisible': ['|', ('state', '=', 'draft'), ('invoice_count','&gt;=',1)]}" context="{'proforma': True}"/>
                     <button name="print_quotation" string="Print" type="object" states="sent,sale" class="o_sale_print"/>
                     <button name="action_quotation_send" string="Send by Email" type="object" states="sent,sale"/>
@@ -245,6 +248,11 @@
                             <field name="partner_id" domain="[('customer','=',True)]" context="{'search_default_customer':1, 'show_address': 1}" options='{"always_reload": True}'/>
                             <field name="partner_invoice_id" groups="sale.group_delivery_invoice_address" context="{'default_type':'invoice'}" options='{"always_reload": True}'/>
                             <field name="partner_shipping_id" groups="sale.group_delivery_invoice_address" context="{'default_type':'delivery'}" options='{"always_reload": True}'/>
+                            <label for="template_id"/>
+                            <div>
+                                <field name="template_id" context="{'company_id': company_id}"
+                                    options="{'no_create': True, 'no_open': True}"/>
+                            </div>
                         </group>
                         <group>
                             <field name="validity_date" attrs="{'invisible': [('state', 'in', ['sale', 'done'])]}"/>
@@ -532,6 +540,8 @@
                                     <field name="user_id"/>
                                     <field name="team_id" options="{'no_create': True}"/>
                                     <field name="client_order_ref"/>
+                                    <field name="require_signature" attrs="{'invisible': [('template_id', '=', False)]}"/>
+                                    <field name="require_payment" attrs="{'invisible': [('template_id', '=', False)]}"/>
                                     <field name="company_id" options="{'no_create': True}" groups="base.group_multi_company"/>
                                     <field name="analytic_account_id" context="{'default_partner_id':partner_invoice_id, 'default_name':name}" attrs="{'readonly': [('invoice_count','!=',0),('state','=','sale')]}" groups="analytic.group_analytic_accounting"/>
                                 </group>
