+++ addons/website_quote/__manifest__.py
@@ -15,7 +15,6 @@
         'views/sale_order_views.xml',
         'views/sale_quote_views.xml',
         'views/website_quote_templates.xml',
-        'views/res_config_settings_views.xml',
         'security/ir.model.access.csv',
     ],
     'demo': [
