+++ addons/website_quote/models/__init__.py
@@ -4,4 +4,3 @@
 from .import sale_order
 from .import sale_quote
 from .import product_template
-from .import res_config_settings
