--- addons/website_quote/models/res_config_settings.py
@@ -1,10 +1,0 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
-from odoo import api, fields, models
-
-
-class ResConfigSettings(models.TransientModel):
-    _inherit = 'res.config.settings'
-
-    default_template_id = fields.Many2one('sale.quote.template', default_model='sale.order', string='Default Template')
