+++ addons/website_quote/models/sale_order.py
@@ -1,31 +1,15 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from datetime import datetime, timedelta
-
-from odoo import api, fields, models, _
+from odoo import api, fields, models
 from odoo.tools.translate import html_translate
-from odoo.exceptions import ValidationError
-
-from werkzeug.urls import url_encode
 
 
 class SaleOrderLine(models.Model):
     _inherit = "sale.order.line"
-    _description = "Sales Order Line"
 
+    # we don't do a related on the product.template quote_description because in this case we want the description to be frozen
     website_description = fields.Html('Line Description', sanitize=False, translate=html_translate)
-
-    # Take the description on the order template if the product is present in it
-    @api.onchange('product_id')
-    def product_id_change(self):
-        domain = super(SaleOrderLine, self).product_id_change()
-        if self.product_id and self.order_id.template_id:
-            for quote_line in self.order_id.template_id.quote_line:
-                if quote_line.product_id == self.product_id:
-                    self.name = quote_line.name
-                    break
-        return domain
 
     @api.model
     def create(self, values):
@@ -40,66 +24,15 @@ class SaleOrderLine(models.Model):
     def _inject_quote_description(self, values):
         values = dict(values or {})
         if not values.get('website_description') and values.get('product_id'):
-            product = self.env['product.product'].browse(values['product_id'])
-            values['website_description'] = product.quote_description or product.website_description
+            product = self.env['product.product'].browse(values['product_id']).with_context(lang=self.order_id.partner_id.lang)
+            values['website_description'] = product.get_quote_description_or_website_description()
         return values
 
 
 class SaleOrder(models.Model):
     _inherit = 'sale.order'
 
-    def _get_default_template(self):
-        template = self.env.ref('website_quote.website_quote_template_default', raise_if_not_found=False)
-        return template and template.active and template or False
-
-    def _get_default_require_signature(self):
-        default_template = self._get_default_template()
-        if default_template:
-            return default_template.require_signature
-        else:
-            return False
-
-    def _get_default_require_payment(self):
-        default_template = self._get_default_template()
-        if default_template:
-            return default_template.require_payment
-        else:
-            return False
-
-    template_id = fields.Many2one(
-        'sale.quote.template', 'Quotation Template',
-        readonly=True,
-        states={'draft': [('readonly', False)], 'sent': [('readonly', False)]},
-        default=_get_default_template)
     website_description = fields.Html('Description', sanitize_attributes=False, translate=html_translate)
-    amount_undiscounted = fields.Float(
-        'Amount Before Discount', compute='_compute_amount_undiscounted', digits=0)
-    require_signature = fields.Boolean('Digital Signature', default=_get_default_require_signature,
-                                       states={'sale': [('readonly', True)], 'done': [('readonly', True)]},
-                                       help='Request a digital signature to the customer in order to confirm orders automatically.')
-    require_payment = fields.Boolean('Electronic Payment', default=_get_default_require_payment,
-                                     states={'sale': [('readonly', True)], 'done': [('readonly', True)]},
-                                     help='Request an electronic payment to the customer in order to confirm orders automatically.')
-
-    @api.multi
-    @api.returns('self', lambda value: value.id)
-    def copy(self, default=None):
-        if self.template_id and self.template_id.number_of_days > 0:
-            default = dict(default or {})
-            default['validity_date'] = fields.Date.to_string(datetime.now() + timedelta(self.template_id.number_of_days))
-        return super(SaleOrder, self).copy(default=default)
-
-    @api.one
-    def _compute_amount_undiscounted(self):
-        total = 0.0
-        for line in self.order_line:
-            total += line.price_subtotal + line.price_unit * ((line.discount or 0.0) / 100.0) * line.product_uom_qty  # why is there a discount in a field named amount_undiscounted ??
-        self.amount_undiscounted = total
-
-    @api.onchange('partner_id')
-    def onchange_partner_id(self):
-        super(SaleOrder, self).onchange_partner_id()
-        self.note = self.template_id.note or self.note
 
     @api.onchange('partner_id')
     def onchange_update_description_lang(self):
@@ -111,142 +44,11 @@ class SaleOrder(models.Model):
 
     @api.onchange('template_id')
     def onchange_template_id(self):
-        if not self.template_id:
-            self.require_signature = False
-            self.require_payment = False
-            return
-        template = self.template_id.with_context(lang=self.partner_id.lang)
-
-        order_lines = [(5, 0, 0)]
-        for line in template.quote_line:
-            data = {
-                'display_type': line.display_type,
-                'name': line.name,
-                'website_description': line.website_description,
-                'state': 'draft',
-            }
-            if line.product_id:
-                discount = 0
-                if self.pricelist_id:
-                    price = self.pricelist_id.with_context(uom=line.product_uom_id.id).get_product_price(line.product_id, 1, False)
-                    if self.pricelist_id.discount_policy == 'without_discount' and line.price_unit:
-                        discount = (line.price_unit - price) / line.price_unit * 100
-                        price = line.price_unit
-
-                else:
-                    price = line.price_unit
-
-                data.update({
-                    'price_unit': price,
-                    'discount': 100 - ((100 - discount) * (100 - line.discount)/100),
-                    'product_uom_qty': line.product_uom_qty,
-                    'product_id': line.product_id.id,
-                    'product_uom': line.product_uom_id.id,
-                    'customer_lead': self._get_customer_lead(line.product_id.product_tmpl_id),
-                })
-                if self.pricelist_id:
-                    data.update(self.env['sale.order.line']._get_purchase_price(self.pricelist_id, line.product_id, line.product_uom_id, fields.Date.context_today(self)))
-            order_lines.append((0, 0, data))
-
-        self.order_line = order_lines
-        self.order_line._compute_tax_id()
-
-        option_lines = []
-        for option in template.options:
-            if self.pricelist_id:
-                price = self.pricelist_id.with_context(uom=option.uom_id.id).get_product_price(option.product_id, 1, False)
-            else:
-                price = option.price_unit
-            data = {
-                'product_id': option.product_id.id,
-                'name': option.name,
-                'quantity': option.quantity,
-                'uom_id': option.uom_id.id,
-                'price_unit': price,
-                'discount': option.discount,
-                'website_description': option.website_description,
-            }
-            option_lines.append((0, 0, data))
-        self.options = option_lines
-
-        if template.number_of_days > 0:
-            self.validity_date = fields.Date.to_string(datetime.now() + timedelta(template.number_of_days))
-
-        self.website_description = template.website_description
-        self.require_signature = template.require_signature
-        self.require_payment = template.require_payment
-
-        if template.note:
-            self.note = template.note
-
-    @api.constrains('template_id', 'require_signature', 'require_payment')
-    def _check_portal_confirmation(self):
-        for order in self.sudo().filtered('template_id'):
-            if not order.require_signature and not order.require_payment:
-                raise ValidationError(_('Please select a confirmation mode in Other Information: Digital Signature, Electronic Payment or both.'))
-
-    @api.multi
-    def open_quotation(self):
-        self.ensure_one()
-        return {
-            'type': 'ir.actions.act_url',
-            'target': 'self',
-            'url': '/quote/%s/%s' % (self.id, self.access_token)
-        }
-
-    @api.multi
-    def get_access_action(self, access_uid=None):
-        """ Instead of the classic form view, redirect to the online quote if it exists. """
-        self.ensure_one()
-        user = access_uid and self.env['res.users'].sudo().browse(access_uid) or self.env.user
-
-        if not self.template_id or (not user.share and not self.env.context.get('force_website')):
-            return super(SaleOrder, self).get_access_action(access_uid)
-        return {
-            'type': 'ir.actions.act_url',
-            'url': '/quote/%s/%s' % (self.id, self.access_token),
-            'target': 'self',
-            'res_id': self.id,
-        }
-
-    def get_mail_url(self):
-        self.ensure_one()
-        if self.state not in ['sale', 'done']:
-            auth_param = url_encode(self.partner_id.signup_get_auth_param()[self.partner_id.id])
-            return '/quote/%s/%s?' % (self.id, self.access_token) + auth_param
-        return super(SaleOrder, self).get_mail_url()
-
-    def get_portal_confirmation_action(self):
-        """ Template override default behavior of pay / sign chosen in sales settings """
+        res = super(SaleOrder, self).onchange_template_id()
         if self.template_id:
-            if self.require_signature and not self.signature:
-                return 'sign'
-            elif self.require_payment:
-                return 'pay'
-            else:
-                return 'none'
-        return super(SaleOrder, self).get_portal_confirmation_action()
-
-    def has_to_be_signed(self):
-        res = super(SaleOrder, self).has_to_be_signed()
-        return self.require_signature if self.template_id else res
-
-    def has_to_be_paid(self):
-        res = super(SaleOrder, self).has_to_be_paid()
-        return self.require_payment if self.template_id else res
-
-    @api.multi
-    def action_confirm(self):
-        res = super(SaleOrder, self).action_confirm()
-        for order in self:
-            if order.template_id and order.template_id.mail_template_id:
-                self.template_id.mail_template_id.send_mail(order.id)
+            template = self.template_id.with_context(lang=self.partner_id.lang)
+            self.website_description = template.website_description
         return res
-
-    @api.multi
-    def _get_payment_type(self):
-        self.ensure_one()
-        return 'form_save' if self.require_payment else 'form'
 
 
 class SaleOrderOption(models.Model):
@@ -256,11 +58,11 @@ class SaleOrderOption(models.Model):
 
     @api.onchange('product_id', 'uom_id')
     def _onchange_product_id(self):
-        if not self.product_id:
-            return
-        product = self.product_id.with_context(lang=self.order_id.partner_id.lang)
-        self.website_description = product.quote_description or product.website_description
-        return super(SaleOrderOption, self)._onchange_product_id()
+        res = super(SaleOrderOption, self)._onchange_product_id()
+        if self.product_id:
+            product = self.product_id.with_context(lang=self.order_id.partner_id.lang)
+            self.website_description = product.get_quote_description_or_website_description()
+        return res
 
     def _get_vals_add_to_order(self):
         vals = super(SaleOrderOption, self)._get_vals_add_to_order()
