+++ addons/website_quote/models/sale_quote.py
@@ -1,108 +1,38 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from odoo import api, fields, models, _
+from odoo import api, fields, models
 from odoo.tools.translate import html_translate
-from odoo.addons import decimal_precision as dp
-from odoo.exceptions import ValidationError, UserError
+
 
 class SaleQuoteTemplate(models.Model):
-    _name = "sale.quote.template"
-    _description = "Sale Quotation Template"
+    _inherit = "sale.quote.template"
 
-    def _get_default_require_signature(self):
-        # A confirmation mode (sign or pay) is mandatory on a quotation template
-        # If none has been activated in the settings, we force 'sign' as confirmation mode
-        if not self.env.user.company_id.portal_confirmation_pay:
-            return True
-        return self.env.user.company_id.portal_confirmation_sign
-
-    def _get_default_require_payment(self):
-        return self.env.user.company_id.portal_confirmation_pay
-
-    name = fields.Char('Quotation Template', required=True)
     website_description = fields.Html('Description', translate=html_translate, sanitize_attributes=False)
-    quote_line = fields.One2many('sale.quote.line', 'quote_id', 'Quotation Template Lines', copy=True)
-    note = fields.Text('Terms and conditions')
-    options = fields.One2many('sale.quote.option', 'template_id', 'Optional Products Lines', copy=True)
-    number_of_days = fields.Integer('Quotation Duration',
-        help='Number of days for the validity date computation of the quotation')
-    require_signature = fields.Boolean('Digital Signature', default=_get_default_require_signature, help='Request a digital signature to the customer in order to confirm orders automatically.')
-    require_payment = fields.Boolean('Electronic Payment', default=_get_default_require_payment,help='Request an electronic payment to the customer in order to confirm orders automatically.')
-    mail_template_id = fields.Many2one(
-        'mail.template', 'Confirmation Mail',
-        domain=[('model', '=', 'sale.order')],
-        help="This e-mail template will be sent on confirmation. Leave empty to send nothing.")
-    active = fields.Boolean(default=True, help="If unchecked, it will allow you to hide the quotation template without removing it.")
-
-    @api.constrains('require_signature', 'require_payment')
-    def _check_confirmation(self):
-        for template in self:
-            if not template.require_signature and not template.require_payment:
-                raise ValidationError(_('Please select a confirmation mode in Confirmation: Digital Signature, Electronic Payment or both.'))
-
-    @api.multi
-    def open_template(self):
-        self.ensure_one()
-        return {
-            'type': 'ir.actions.act_url',
-            'target': 'self',
-            'url': '/quote/template/%d' % self.id
-        }
 
 
 class SaleQuoteLine(models.Model):
-    _name = "sale.quote.line"
-    _description = "Quotation Template Lines"
-    _order = 'quote_id, sequence, id'
+    _inherit = "sale.quote.line"
 
-    sequence = fields.Integer('Sequence', help="Gives the sequence order when displaying a list of sale quote lines.",
-        default=10)
-    quote_id = fields.Many2one('sale.quote.template', 'Quotation Template Reference', required=True,
-        ondelete='cascade', index=True)
-    name = fields.Text('Description', required=True, translate=True)
-    product_id = fields.Many2one('product.product', 'Product', domain=[('sale_ok', '=', True)])
+    # this website_description is related because we want to update the product.template quote_description
+    # when we update the quote.line website_description with the website editor
     website_description = fields.Html('Line Description', related='product_id.product_tmpl_id.quote_description',
         translate=html_translate)
-    price_unit = fields.Float('Unit Price', required=True, digits=dp.get_precision('Product Price'))
-    discount = fields.Float('Discount (%)', digits=dp.get_precision('Discount'), default=0.0)
-    product_uom_qty = fields.Float('Quantity', required=True, digits=dp.get_precision('Product UoS'), default=1)
-    product_uom_id = fields.Many2one('uom.uom', 'Unit of Measure')
-
-    display_type = fields.Selection([
-        ('line_section', "Section"),
-        ('line_note', "Note")], default=False, help="Technical field for UX purpose.")
 
     @api.onchange('product_id')
     def _onchange_product_id(self):
         self.ensure_one()
         if self.product_id:
-            name = self.product_id.name_get()[0][1]
-            if self.product_id.description_sale:
-                name += '\n' + self.product_id.description_sale
-            self.name = name
-            self.price_unit = self.product_id.lst_price
-            self.product_uom_id = self.product_id.uom_id.id
-            self.website_description = self.product_id.quote_description or self.product_id.website_description or ''
-            domain = {'product_uom_id': [('category_id', '=', self.product_id.uom_id.category_id.id)]}
-            return {'domain': domain}
-
-    @api.onchange('product_uom_id')
-    def _onchange_product_uom(self):
-        if self.product_id and self.product_uom_id:
-            self.price_unit = self.product_id.uom_id._compute_price(self.product_id.lst_price, self.product_uom_id)
+            self.website_description = self.product_id.get_quote_description_or_website_description()
+        return super(SaleQuoteLine, self)._onchange_product_id()
 
     @api.model
     def create(self, values):
-        if values.get('display_type', self.default_get(['display_type'])['display_type']):
-            values.update(product_id=False, price_unit=0, product_uom_qty=0, product_uom_id=False)
         values = self._inject_quote_description(values)
         return super(SaleQuoteLine, self).create(values)
 
     @api.multi
     def write(self, values):
-        if 'display_type' in values and self.filtered(lambda line: line.display_type != values.get('display_type')):
-            raise UserError("You cannot change the type of a sale quote line. Instead you should delete the current line and create a new line of the proper type.")
         values = self._inject_quote_description(values)
         return super(SaleQuoteLine, self).write(values)
 
@@ -110,52 +40,28 @@ class SaleQuoteLine(models.Model):
         values = dict(values or {})
         if not values.get('website_description') and values.get('product_id'):
             product = self.env['product.product'].browse(values['product_id'])
-            values['website_description'] = product.quote_description or product.website_description or ''
+            values.update(website_description=product.get_quote_description_or_website_description())
         return values
 
-    _sql_constraints = [
-        ('accountable_product_id_required',
-            "CHECK(display_type IS NOT NULL OR (product_id IS NOT NULL AND product_uom_id IS NOT NULL))",
-            "Missing required product and UoM on accountable sale quote line."),
-
-        ('non_accountable_fields_null',
-            "CHECK(display_type IS NULL OR (product_id IS NULL AND price_unit = 0 AND product_uom_qty = 0 AND product_uom_id IS NULL))",
-            "Forbidden product, unit price, quantity, and UoM on non-accountable sale quote line"),
-    ]
+    def _get_vals_to_apply_template(self):
+        vals = super(SaleQuoteLine, self)._get_vals_to_apply_template()
+        vals.update(website_description=self.website_description)
+        return vals
 
 
 class SaleQuoteOption(models.Model):
-    _name = "sale.quote.option"
-    _description = "Quotation Option"
+    _inherit = "sale.quote.option"
 
-    template_id = fields.Many2one('sale.quote.template', 'Quotation Template Reference', ondelete='cascade',
-        index=True, required=True)
-    name = fields.Text('Description', required=True, translate=True)
-    product_id = fields.Many2one('product.product', 'Product', domain=[('sale_ok', '=', True)], required=True)
+    # why for options we don't modify the one from the product?
     website_description = fields.Html('Option Description', translate=html_translate, sanitize_attributes=False)
-    price_unit = fields.Float('Unit Price', required=True, digits=dp.get_precision('Product Price'))
-    discount = fields.Float('Discount (%)', digits=dp.get_precision('Discount'))
-    uom_id = fields.Many2one('uom.uom', 'Unit of Measure ', required=True)
-    quantity = fields.Float('Quantity', required=True, digits=dp.get_precision('Product UoS'), default=1)
 
     @api.onchange('product_id')
     def _onchange_product_id(self):
-        if not self.product_id:
-            return
-        product = self.product_id
-        self.price_unit = product.list_price
-        self.website_description = product.product_tmpl_id.quote_description
-        self.name = product.name
-        self.uom_id = product.uom_id
-        domain = {'uom_id': [('category_id', '=', self.product_id.uom_id.category_id.id)]}
-        return {'domain': domain}
+        if self.product_id:
+            self.website_description = self.product_id.get_quote_description_or_website_description()
+        return super(SaleQuoteOption, self)._onchange_product_id()
 
-    @api.onchange('uom_id')
-    def _onchange_product_uom(self):
-        if not self.product_id:
-            return
-        if not self.uom_id:
-            self.price_unit = 0.0
-            return
-        if self.uom_id.id != self.product_id.uom_id.id:
-            self.price_unit = self.product_id.uom_id._compute_price(self.price_unit, self.uom_id)
+    def _get_vals_to_apply_template(self):
+        vals = super(SaleQuoteOption, self)._get_vals_to_apply_template()
+        vals.update(website_description=self.website_description)
+        return vals
