+++ addons/website_quote/security/ir.model.access.csv
@@ -1,7 +1,1 @@
 id,name,model_id:id,group_id:id,perm_read,perm_write,perm_create,perm_unlink
-access_sale_quote_template,sale.quote.template,model_sale_quote_template,sales_team.group_sale_salesman,1,0,0,0
-access_sale_quote_template_manager,sale.quote.template,model_sale_quote_template,sales_team.group_sale_manager,1,1,1,1
-access_sale_quote_line,sale.quote.line,model_sale_quote_line,sales_team.group_sale_salesman,1,0,0,0
-access_sale_quote_line_manager,sale.quote.line,model_sale_quote_line,sales_team.group_sale_manager,1,1,1,1
-access_sale_quote_option,sale.quote.option,model_sale_quote_option,sales_team.group_sale_salesman,1,0,0,0
-access_sale_quote_option_manager,sale.quote.option,model_sale_quote_option,sales_team.group_sale_manager,1,1,1,1
