+++ addons/website_quote/views/website_quote_templates.xml
@@ -669,7 +669,7 @@
                     <div class="col-lg-9">
                         <div class="row mb8">
                             <div t-ignore="true" class="css_editable_mode_hidden offset-lg-11">
-                              <a role="button" class="btn btn-info d-print-none" t-att-href="'/web#return_label=Website&amp;model=%s&amp;id=%s&amp;action=%s&amp;view_type=form' % (template._name, template.id, request.env.ref('website_quote.action_sale_quotation_template').id)">Back</a>
+                              <a role="button" class="btn btn-info d-print-none" t-att-href="'/web#return_label=Website&amp;model=%s&amp;id=%s&amp;action=%s&amp;view_type=form' % (template._name, template.id, request.env.ref('sale.action_sale_quotation_template').id)">Back</a>
                             </div>
                         </div>
                         <div class="alert alert-info alert-dismissable" t-ignore="True" role="status">
